Rename touched files in line with agreed standards
