+++ resources-domain/resources-service/src/resource/sql/resourceGetSql.ts
