+++ resources-domain/resources-service/tests/unit/resource/resourcesRoutesV0.test.ts
@@ -17,14 +17,14 @@
 import { ReferrableResource } from '@tech-matters/types';
 
 import { Request, Response, Router } from 'express';
-import resourceRoutes from '../../../src/resource/resource-routes-v0';
-import { resourceModel } from '../../../src/resource/resource-model';
+import resourceRoutes from '../../../src/resource/resourceRoutesV0';
+import { resourceService } from '../../../src/resource/resourceService';
 
 jest.mock('express', () => ({
   Router: jest.fn(),
 }));
 
-jest.mock('../../../src/resource/resource-model', () => ({
+jest.mock('../../../src/resource/resourceService', () => ({
   resourceModel: jest.fn(),
 }));
 
@@ -33,7 +33,7 @@ const mockSearchResources: jest.Mock<Promise<{
   results: ReferrableResource[];
 }>> = jest.fn();
 
-(<jest.Mock>resourceModel).mockReturnValue({
+(<jest.Mock>resourceService).mockReturnValue({
   searchResourcesByName: mockSearchResources,
 });
